requirements.txt
Flask
Flask-SQLAlchemy
gunicorn
pytest


